backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# Database (MySQL optional for testing)
# mysqlclient==2.2.0

# Caching (Redis optional for testing)
# redis==5.0.1
# django-redis==5.4.0

# AI/ML (will be added in Task 8)
# spacy==3.7.2
# scikit-learn==1.3.2

# HTTP requests
requests==2.31.0

# Configuration
python-decouple==3.6

# Production (optional for testing)
# celery==5.3.4
# gunicorn==21.2.0
# whitenoise==6.6.0

# Development
django-extensions==3.2.3